Drop side effects and dropped slots for nodes and their dumping
This are the first exploratory isolation tests to write in an attempt
to learn the entire original implementation of this and they focus on
the node class initialization and dumping.

Remove unused slots like q_dict and add type hints to all methods
of the node class to wrap up its refactoring.