Add initial tokens submodule and representation in rust
This includes a rust-based enum and implementations for the python
token classes as well as a tokens submodule to hold all current
token functionality now to be migrated from the python module.

The initially introduced enum implementing display and equality
traits have newly added test checks on the rust side.