Add initial tokens submodule and representation in rust
This includes a rust-based struct and implementations for the python
token class as well as a tokens submodule to hold all previous token
functionality now to be migrated to rust.

The initially introduced struct implementing comparable traits has
all necessary checks for equivalence relations and str/repr but is
not yet ready to be used within the python codebase.