Create initial frontend and backend CI workflows
The frontend was created by maturin with support for all relevant
OS and architecture variants and focuses on building python wheels.
The backend focuses on building the rust code and running tests for
the rust codebase. The original CI workflow focuses on isolation and
integration tests from actual usage as before but now needs a maturin
build and installation step in a virtual environment in order to
provide the shared object file and achieve this (thus build and
dependencies even before the isolation tests).

The workflows will run on any and all branches by default to greater
flexibility regarding any development model which differs from the
default that would only focus on a main branch.