Document, annotate, and test all public methods of the parser class
This provides exploratory isolation tests for the parser class
when it comes to all of its public methods but without diving
into the main use cases.

Besides a few argument renames and method sorting and renames
for clarity, one small fix was performed to provide a proper
default node for the previous `multiply_join` method which does
not imply real multiplication (meet) but addition (join) of
filtered dictionaries.