Migrate from setup.py to pyproject.toml packaging
The maturin build system requires the newer pyproject.toml based
packaging for modern python projects so this is a perfect time
to perform the switch while still providing avocado plugin entry
points for the rust-enabled variant of the CartConf project.

The initially added pyproject.toml file has now been updated with
all the information that was previously needed for the integration
with the avocado plugins ecosystem.

This was also needed in order to be able to pass the second (and
last) integration test in the CI.