Perform code optimizations regarding string concatenation
Use lists of characters or strings joined once at the end instead
of more performance-intensive string concatenation at each step.

The node dumping will also not include a newline character at the
end of the node representation allowing for the caller to decide
how to integrate the dump into the output.