Refactor core private methods of the parser decoupling dependencies
We separate the various major conditions of the main `_parse` entry
point to the core parsing functionality into private methods with
as little coupling as possible.

Returning an empty dictionary upon reset is an awkward way to do
a dictionary reset in a manner that still respects the high previous
coupling for passing acceptance testing. Now however we can take a
next step towards simplifying the API and functional separation of
the core parsing method so copy and clear the pre-dictionary instead
of returning an empty one.

Similar simplification was done for the variants parsing reordering
and decoupling further from the variant metadata.