Migrate the regex value substitution to a public rust tokens function
This substitution function is used by all operator tokens but in
turn uses a drop_suffixes call which we will temporarily call in
reverse (python call from rust) before we could decouple further.
This will introduce a slow down due to initializing python each
time we perform a value substitution by a dictionary's keys but
such slowdown remains minor and hopefully temporary.

The substitution python function is also fully tested in isolation
on the rust side in order to be understood better and to validate
this translation between both languages (but for this commit still
needs python virtual environment with the cartconf wheel installed).

Within the new implementation we could not automatically extract
the hashmap from the python dictionary as it also contains lists
for values of "deps" and other keys and is generally unclean when
it comes to a proper value type for the parameters dictionary. We
thus extract a rust hashmap only for string values by ourselves
within the lifetime of the python call.