Duplicate the suffix dropping functionality on the Rust side
Duplicate reserved keys on the rust side temporarily just like the
suffix dropping function until we migrate more code.

The python unit tests are still somewhat slower despite the fact
that reverse rust2python calls have now been dropped but the
expectation is that it will speed up again once we reduce the
python-rust crosstalk bandwidth further.

We also still do a lot of unwrapping in this function similarly to
the regex substitution function before we fully migrate it when
we could potentially switch to returning a result.