Restore "slow" parsing support for operator tokens
It seems that the operator fast mode was always enabled in the
manual testing and practical use of the parser during contributions
which introduced a regression in no longer supporting normal parsing
of operator tokens. In particular, arbitrary characters must be
allowed in the values after the operator identifiers and regex
characters must be allowed in the keys. If fast mode was truly meant
as the only operator matching there would be no operator handling
in the stricter mode so this must indeed be a regression. Oddly,
removing fast mode from the parse() method makes overall parsing
in unit test slightly faster.

One complication from the current fix is that regex capable operators
require a space to separate the identifier regex key from the value.
This makes sense as a requirement since "abc??=a" will be ambiguous
and only "abc? ?= a" will be unambiguous anyway.

Most of the effort here is actually to avoid complex filters and
default variants respectively using forms like "(a=b)" and "[a=b]"
and to correctly separate regex characters in the operator key like
"a(_[cd]+)* ?= b" from the former despite overlapping characters.

This is also best done separately from the lexer migration where it
will have to be translated next. It works according to all test
coverage we have so any further issues will be caught by regression
tests. While I am still not sure this will work e.g. with filters or
all conditionals or various only/no variants, a bigger wish is that
only the affected chars should be treated and skips should only be
done on double ?? or double .. instead of skipping everything for all
others chars too.