Use simpler and more meaningful rust-side check_token signature
The trick of returning the same token and its easily extractable
type is local to just the filter parsing and as such must be moved
back there as localized syntactic sugar.

This was also an opportunity to factor out the token validation
for next token retrieval as well. Needless to say, such a check
should also raise a lexer error.