Migrate the first and simplest (pre-dict) application method
This will require some python object tolerance in rust as well as
tolerance towards the side-effect-heavy pre-dict modifications but
all of them are meant to be temporary until the migration progresses.

Also prepare operator (incl. deletion) arguments with static methods
where possible.

We do not provide lexer auto-conversion and instead treat it within
python since we would need cloning trait not just for it but also
for the reader enum.