Introduce content step to simplify the previous complex type
This makes clippy as a rust linter happier and improves the overall
readability of the node implementation as the overall tuple and its
various parts are all named thus always dumped with explanations.

A struct for failed case is still on table but this isolated the
linter macros well enough to just a single case.