Migrate the apply-include with first ever reverse parser call
This is the first step towards dealing with the multi-function
recursion between parsing and apply methods with the rest of the
translation following the guidelines of few previous methods.

The same migration step is also a first demonstration of rust-side
lexer creation that is then passed back to the python side.