Migrate the apply variants method (plural)
This is the first of the two last (variantization) apply methods
and we could not batch the two despite one being practically a
consumer for the other because the node label type would need more
coherent hashing and equality support between the python and rust
side.

While this method didn't have inverse _parse calls it did make use
of lots of inverse lexer calls as well as inverse parser errors.
A lot ot not all of these can be simplified later with a reduced
cross-talk and thus fewer type conversions.