Use consistent types in previously less understood variant defaults
The meta type has been underdocumented, undertested, and underused
so the complete type discovered via the translation into a statically
typed language can now be simplified together with a few other cases
of inconsistent default variant types. The choice of simpler type is
a fairly standard for python list of strings in both cases.