Make the label type displayable and hashable also on the rust side
This is needed if we want to employ hash sets on the rust side in
addition to the python side and will be useful for future use cases
of the labels' hash functionality also within rust. We could not
auto-derive it due to our custom implementation of the PartialEq
trait and providing a custom implementation fits better with the
custom hashing functionality already present for the python side.
The `Eq` trait is an additional requirement for iterators that
could be auto-derived in comparison.

The display trait also could not be auto-derived just like the debug
trait and is needed to be able to easily convert labels to strings.