Migrate the apply variant method (singular)
This is the second of the two last (variantization) apply methods
which needed two extra parser attributes to be passed along as
arguments for local use so that the node apply methods functionality
is complete. The method also constructs new nodes and adds the current
one as a child which could lead to additional cloning we should look
into simplifying with more migration.