Use pointer size integers consistently
The unsigned pointer size type would be most appropriate but we
have to leave a TODO about avoiding -1 as either indent or a line
number first and on the parser level it is currently safest to just
rely on signed pointer size and avoid further usize conversions.

The preference is on architecture adaptive size instead of fixed
32 bits which despite consistent behavior across architectures
introduces hardcoding and possibility of harder eventual migration.