Reduce higher order python arg types to default rust struct variants
Since Rust is more limited in terms of introspection (something that
is easier to achieve with dynamically typed languages like python),
we cannot pass along the previous python types (higher order sort)
as actual enum variant arguments to the lexer. Instead, provide
default instances that can then be type matched in a more idiomatic
way within rust. We could take further advantage of these if we later
on decide to match particular data values and thus instances in the
lexer.

Leave some todos to focus this commit better:

1) Some tokens were have no defaults due to identifier-based inversion.
2) Some reuse of default tokens (instead of recreation each time) is
   highly recommended.