Use param key and val enums for python pre_dict data polytypes
The key should handle both a string and a tuple of an arbitrary
length and the value should handle both a string and a list of
strings. Use these in a rust hashmap as parsing pre-dict with
suffixable parameter keys and extended values.

Instead of cloning and then clearing a complete hashmap during
pre-dict application, we rely on `mem::take` to save at least
some performance. The entire pre-dict resetting side effect
should instead be dropped by simply consuming such pre-dicts
during application in order to reduce complexity. However, we
will only do actual refactoring once the migration is complete
and fully validated. Similarly we also use `mem::replace`.

While the pre-dict application was simplified entirely out of
python context, the rest of the apply methods with py context
were kept around (even when the python object is not used) until
each recursive call is freed entirely from the need of such a py
object and therefore closure.

Substitution now borrows and is thus unavailable for python code.