Move variant meta python dictionary to a mutable hashmap reference
This requires (1) getting mutable default item of that mutable
referenced hashmap and (2) providing a mutable reference to an
empty vector instead of none option in order to be able to reuse
the meta default values inside and outside the variant application
loop. The latter means that instead of separate treatment of none
option and an empty vector we end up with a simpler unified treatment
of both as a mutable reference to an empty vector we won't modify
anyway. The error at the end of the loop is only returned if the
list is nonempty which will also immediately imply some (rather
than none) option so this should be safe enough yet simpler.