Standardize missing-include error to a regular parser error
The error message and details order is inverted in the exceptions
module but for no apparent reason. Instead of a single location
use and fixed already known error message but a separate class,
we now rely on a regular parser error with the same error message
while deprecating the python exceptions module.

Now that the last python crack into a newly pure rust parsing core
has been removed, we can only have parser and lexer exceptions as
possible errors and can thus standardize the line argument handling
of these in all return instances.