Add filter joining to the explicit recursive state machine
Integrate the inner implicit python generator state machine into
the outer explicit dictionary getter state machine by expanding
the latter with joins as simpler stateful data thus making it
possible to migrate all dictionary getters and the overall pre-dict
to rust. Additional decoupled (sub-)pre-dicts are necessary to
remove part of the statefulness and start multiple "states" from
the same initial pre-join state via a shallow copy. Furthermore,
dicts cache is added for extra speedup to save on processing such
pre-dicts as much as possible.

While part of the previous join recursion is now skipped and cached,
new joining still recurses and so do the plain getter (to handle
next) as well as transitions between the join and plain getters.
At present, the dict getter arbiter skips straight to the final
pre-dict step for cases of joined dicts and traversers each child
for the case of plain dicts (and finished join dicts) while more
of the join node preprocessing is cached for later reuse.

Extra coverage was needed in order to validate this translation of
implementations which was severly missing so far.