Perform content-independent parser node comparison
This allows for proper use of the content swapping method for
operations that need temporarily different content, e.g. joining
that needs all join filters to be dropped and replaced with one
only filter at a time. Such cases did not have reliable node
comparison and the only reason we didn't have any trouble is that
no functionality depended on node comparison so far (which will
be the case when moving to an explicit recursive state machine).

This also improves the performance of node comparison since the
default partial equality implementation in rust would compare all
fields of which the node has the most.

The ID-s attached to the nodes will only depend on one constructor
call after which we can swap the node content as much as we like.